myapi/requirements.txt
Django>=4.0,<5.0
djangorestframework
django-cors-headers
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
boto3==1.34.145
botocore==1.34.145
pandas==2.2.2
plotly==5.22.0
SQLAlchemy==2.0.31
mysql-connector==2.2.9
firebase-admin==5.3.0
environ
django-environ